@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 class DetailsScreen {
   const DetailsScreen({
     this.appBar,
-    @required this.body,
+    required this.body,
   });
 
   final Widget body;
-  final PreferredSizeWidget appBar;
+  final PreferredSizeWidget? appBar;
 }
