@@ -5,16 +5,16 @@ import '../three_column_navigation.dart';
 
 class SectionList extends StatelessWidget {
   const SectionList({
-    Key key,
-    @required this.controller,
-    @required MainSection section,
-    @required int listIndex,
-    @required this.listTap,
-  })  : _section = section,
+    Key? key,
+    required this.controller,
+    required MainSection section,
+    required int listIndex,
+    required this.listTap,
+  })   : _section = section,
         _listIndex = listIndex,
         super(key: key);
 
-  final AutoScrollController controller;
+  final AutoScrollController? controller;
   final MainSection _section;
   final int _listIndex;
   final ValueChanged<int> listTap;
@@ -26,12 +26,12 @@ class SectionList extends StatelessWidget {
       itemBuilder: (context, index) {
         final _item = _section.itemBuilder(context, index, _listIndex == index);
         return AutoScrollTag(
-          controller: controller,
+          controller: controller!,
           key: ValueKey(index),
           index: index,
           child: GestureDetector(
-            child: _item,
             onTap: () => listTap(index),
+            child: _item,
           ),
         );
       },
