@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 class ResponsiveScaffold extends StatelessWidget {
   const ResponsiveScaffold({
+    Key? key,
     this.scaffoldKey,
     this.drawer,
     this.endDrawer,
@@ -14,27 +15,27 @@ class ResponsiveScaffold extends StatelessWidget {
     this.kTabletBreakpoint = 720.0,
     this.kDesktopBreakpoint = 1440.0,
     this.appBarElevation,
-  });
+  }) : super(key: key);
 
-  final Widget drawer, endDrawer;
+  final Widget? drawer, endDrawer;
 
-  final Widget title;
+  final Widget? title;
 
-  final Widget body;
+  final Widget? body;
 
-  final Widget trailing;
+  final Widget? trailing;
 
-  final Widget floatingActionButton;
+  final Widget? floatingActionButton;
 
-  final kTabletBreakpoint;
-  final kDesktopBreakpoint;
-  final _drawerWidth = 304.0;
+  final double kTabletBreakpoint;
+  final double kDesktopBreakpoint;
+  double get _drawerWidth => 304.0;
 
-  final IconData menuIcon, endIcon;
+  final IconData? menuIcon, endIcon;
 
-  final double appBarElevation;
+  final double? appBarElevation;
 
-  final Key scaffoldKey;
+  final Key? scaffoldKey;
 
   @override
   Widget build(BuildContext context) {
@@ -51,7 +52,7 @@ class ResponsiveScaffold extends StatelessWidget {
                         width: _drawerWidth,
                         child: Drawer(
                           child: SafeArea(
-                            child: drawer,
+                            child: drawer!,
                           ),
                         ),
                       ),
@@ -65,7 +66,7 @@ class ResponsiveScaffold extends StatelessWidget {
                           title: title,
                           actions: <Widget>[
                             if (trailing != null) ...[
-                              trailing,
+                              trailing!,
                             ],
                           ],
                         ),
@@ -75,12 +76,12 @@ class ResponsiveScaffold extends StatelessWidget {
                               child: body ?? Container(),
                             ),
                             if (endDrawer != null) ...[
-                              Container(
+                              SizedBox(
                                 width: _drawerWidth,
                                 child: Drawer(
                                   elevation: 3.0,
                                   child: SafeArea(
-                                    child: endDrawer,
+                                    child: endDrawer!,
                                   ),
                                 ),
                               ),
@@ -95,7 +96,7 @@ class ResponsiveScaffold extends StatelessWidget {
                   Positioned(
                     top: 100.0,
                     left: _drawerWidth - 30,
-                    child: floatingActionButton,
+                    child: floatingActionButton!,
                   )
                 ],
               ],
@@ -109,7 +110,7 @@ class ResponsiveScaffold extends StatelessWidget {
                 ? null
                 : Drawer(
                     child: SafeArea(
-                      child: drawer,
+                      child: drawer!,
                     ),
                   ),
             appBar: AppBar(
@@ -119,7 +120,7 @@ class ResponsiveScaffold extends StatelessWidget {
               leading: _MenuButton(iconData: menuIcon),
               actions: <Widget>[
                 if (trailing != null) ...[
-                  trailing,
+                  trailing!,
                 ],
               ],
             ),
@@ -134,12 +135,12 @@ class ResponsiveScaffold extends StatelessWidget {
                         child: body ?? Container(),
                       ),
                       if (endDrawer != null) ...[
-                        Container(
+                        SizedBox(
                           width: _drawerWidth,
                           child: Drawer(
                             elevation: 3.0,
                             child: SafeArea(
-                              child: endDrawer,
+                              child: endDrawer!,
                             ),
                           ),
                         ),
@@ -150,7 +151,7 @@ class ResponsiveScaffold extends StatelessWidget {
                     Positioned(
                       top: 10.0,
                       left: 10.0,
-                      child: floatingActionButton,
+                      child: floatingActionButton!,
                     )
                   ],
                 ],
@@ -164,14 +165,14 @@ class ResponsiveScaffold extends StatelessWidget {
               ? null
               : Drawer(
                   child: SafeArea(
-                    child: drawer,
+                    child: drawer!,
                   ),
                 ),
           endDrawer: endDrawer == null
               ? null
               : Drawer(
                   child: SafeArea(
-                    child: endDrawer,
+                    child: endDrawer!,
                   ),
                 ),
           appBar: AppBar(
@@ -181,7 +182,7 @@ class ResponsiveScaffold extends StatelessWidget {
             title: title,
             actions: <Widget>[
               if (trailing != null) ...[
-                trailing,
+                trailing!,
               ],
               if (endDrawer != null) ...[
                 _OptionsButton(iconData: endIcon),
@@ -199,11 +200,11 @@ class ResponsiveScaffold extends StatelessWidget {
 
 class _OptionsButton extends StatelessWidget {
   const _OptionsButton({
-    Key key,
-    @required this.iconData,
+    Key? key,
+    required this.iconData,
   }) : super(key: key);
 
-  final IconData iconData;
+  final IconData? iconData;
 
   @override
   Widget build(BuildContext context) {
@@ -218,11 +219,11 @@ class _OptionsButton extends StatelessWidget {
 
 class _MenuButton extends StatelessWidget {
   const _MenuButton({
-    Key key,
-    @required this.iconData,
+    Key? key,
+    required this.iconData,
   }) : super(key: key);
 
-  final IconData iconData;
+  final IconData? iconData;
 
   @override
   Widget build(BuildContext context) {
